requirements.txt
Django==3.1.5
psycopg2==2.8.6
djangorestframework==3.12.2
requests==2.25.1
djangorestframework-simplejwt==4.6.0